Add Composer & PSR Namespace
